+++ firebase-vscode/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Updated internal firebase-tools dependency to 13.20.2
 - Updated default debug-log output to .firebase/logs directory
+- [Fixed] Fixed an issue where emulator startup would hang
 
 ## 0.10.1